+++ packages/components/select/src/use-select.ts
@@ -1,36 +1,27 @@
 import type {SelectSlots, SelectVariantProps, SlotsToClasses} from "@heroui/theme";
 import type {HiddenSelectProps} from "./hidden-select";
+import type {DOMAttributes, HTMLHeroUIProps, PropGetter, SharedSelection} from "@heroui/system";
+import type {ReactRef} from "@heroui/react-utils";
+import type {Key, ReactNode} from "react";
+import type {ListboxProps} from "@heroui/listbox";
+import type {PopoverProps} from "@heroui/popover";
+import type {ScrollShadowProps} from "@heroui/scroll-shadow";
+import type {MultiSelectProps, MultiSelectState} from "@heroui/use-aria-multiselect";
+import type {SpinnerProps} from "@heroui/spinner";
+import type {CollectionChildren, ValidationError} from "@react-types/shared";
 
-import {
-  DOMAttributes,
-  HTMLHeroUIProps,
-  mapPropsVariants,
-  PropGetter,
-  SharedSelection,
-  useLabelPlacement,
-  useProviderContext,
-} from "@heroui/system";
+import {mapPropsVariants, useLabelPlacement, useProviderContext} from "@heroui/system";
 import {select} from "@heroui/theme";
-import {ReactRef, useDOMRef, filterDOMProps} from "@heroui/react-utils";
-import {useMemo, useCallback, useRef, Key, ReactNode, useEffect} from "react";
-import {ListboxProps} from "@heroui/listbox";
+import {useDOMRef, filterDOMProps} from "@heroui/react-utils";
+import {useMemo, useCallback, useRef, useEffect} from "react";
 import {useAriaButton} from "@heroui/use-aria-button";
 import {useFocusRing} from "@react-aria/focus";
 import {clsx, dataAttr, objectToDeps} from "@heroui/shared-utils";
 import {mergeProps} from "@react-aria/utils";
 import {useHover} from "@react-aria/interactions";
-import {PopoverProps} from "@heroui/popover";
-import {ScrollShadowProps} from "@heroui/scroll-shadow";
-import {
-  MultiSelectProps,
-  MultiSelectState,
-  useMultiSelect,
-  useMultiSelectState,
-} from "@heroui/use-aria-multiselect";
-import {SpinnerProps} from "@heroui/spinner";
+import {useMultiSelect, useMultiSelectState} from "@heroui/use-aria-multiselect";
 import {useSafeLayoutEffect} from "@heroui/use-safe-layout-effect";
 import {ariaShouldCloseOnInteractOutside} from "@heroui/aria-utils";
-import {CollectionChildren, ValidationError} from "@react-types/shared";
 import {FormContext, useSlottedContext} from "@heroui/form";
 import {usePreventScroll} from "@react-aria/overlays";
 
@@ -354,10 +345,7 @@ export function useSelect<T extends object>(originalProps: UseSelectProps<T>) {
   });
 
   const hasPlaceholder = !!placeholder;
-  const shouldLabelBeOutside =
-    labelPlacement === "outside-left" ||
-    (labelPlacement === "outside" &&
-      (!(hasPlaceholder || !!description) || !!originalProps.isMultiline));
+  const shouldLabelBeOutside = labelPlacement === "outside-left" || labelPlacement === "outside";
   const shouldLabelBeInside = labelPlacement === "inside";
   const isOutsideLeft = labelPlacement === "outside-left";
 
@@ -370,6 +358,7 @@ export function useSelect<T extends object>(originalProps: UseSelectProps<T>) {
     !!originalProps.isMultiline;
   const hasValue = !!state.selectedItems?.length;
   const hasLabel = !!label;
+  const hasLabelOutside = hasLabel && (isOutsideLeft || (shouldLabelBeOutside && hasPlaceholder));
 
   const baseStyles = clsx(classNames?.base, className);
 
@@ -383,25 +372,6 @@ export function useSelect<T extends object>(originalProps: UseSelectProps<T>) {
       }),
     [objectToDeps(variantProps), isInvalid, labelPlacement, disableAnimation],
   );
-
-  // scroll the listbox to the selected item
-  useEffect(() => {
-    if (state.isOpen && popoverRef.current && listBoxRef.current) {
-      let selectedItem = listBoxRef.current.querySelector("[aria-selected=true] [data-label=true]");
-      let scrollShadow = scrollShadowRef.current;
-
-      // scroll the listbox to the selected item
-      if (selectedItem && scrollShadow && selectedItem.parentElement) {
-        let scrollShadowRect = scrollShadow?.getBoundingClientRect();
-        let scrollShadowHeight = scrollShadowRect.height;
-
-        scrollShadow.scrollTop =
-          selectedItem.parentElement.offsetTop -
-          scrollShadowHeight / 2 +
-          selectedItem.parentElement.clientHeight / 2;
-      }
-    }
-  }, [state.isOpen, disableAnimation]);
 
   usePreventScroll({
     isDisabled: !state.isOpen,
@@ -432,12 +402,13 @@ export function useSelect<T extends object>(originalProps: UseSelectProps<T>) {
       "data-has-label": dataAttr(hasLabel),
       "data-has-helper": dataAttr(hasHelper),
       "data-invalid": dataAttr(isInvalid),
+      "data-has-label-outside": dataAttr(hasLabelOutside),
       className: slots.base({
         class: clsx(baseStyles, props.className),
       }),
       ...props,
     }),
-    [slots, hasHelper, hasValue, hasLabel, isFilled, baseStyles],
+    [slots, hasHelper, hasValue, hasLabel, hasLabelOutside, isFilled, baseStyles],
   );
 
   const getTriggerProps: PropGetter = useCallback(
@@ -496,7 +467,7 @@ export function useSelect<T extends object>(originalProps: UseSelectProps<T>) {
         form: originalProps?.form,
         onChange,
         ...props,
-      } as HiddenSelectProps<T>),
+      }) as HiddenSelectProps<T>,
     [
       state,
       selectionMode,