+++ packages/components/toast/src/toast-provider.tsx
@@ -1,8 +1,14 @@
-import {ToastOptions, ToastQueue, useToastQueue} from "@react-stately/toast";
-import {useProviderContext} from "@heroui/system";
+import type {ToastOptions} from "@react-stately/toast";
+import type {RegionProps} from "./toast-region";
+import type {ToastProps, ToastPlacement} from "./use-toast";
 
-import {RegionProps, ToastRegion} from "./toast-region";
-import {ToastProps, ToastPlacement} from "./use-toast";
+import {ToastQueue, useToastQueue} from "@react-stately/toast";
+import {useProviderContext} from "@heroui/system";
+import {AnimatePresence, LazyMotion} from "framer-motion";
+
+import {ToastRegion} from "./toast-region";
+
+const loadFeatures = () => import("framer-motion").then((res) => res.domMax);
 
 let globalToastQueue: ToastQueue<ToastProps> | null = null;
 
@@ -37,20 +43,22 @@ export const ToastProvider = ({
   const globalContext = useProviderContext();
   const disableAnimation = disableAnimationProp ?? globalContext?.disableAnimation ?? false;
 
-  if (toastQueue.visibleToasts.length == 0) {
-    return null;
-  }
-
   return (
-    <ToastRegion
-      disableAnimation={disableAnimation}
-      maxVisibleToasts={maxVisibleToasts}
-      placement={placement}
-      toastOffset={toastOffset}
-      toastProps={toastProps}
-      toastQueue={toastQueue}
-      {...regionProps}
-    />
+    <LazyMotion features={loadFeatures}>
+      <AnimatePresence>
+        {toastQueue.visibleToasts.length > 0 ? (
+          <ToastRegion
+            disableAnimation={disableAnimation}
+            maxVisibleToasts={maxVisibleToasts}
+            placement={placement}
+            toastOffset={toastOffset}
+            toastProps={toastProps}
+            toastQueue={toastQueue}
+            {...regionProps}
+          />
+        ) : null}
+      </AnimatePresence>
+    </LazyMotion>
   );
 };